move extractParam to its own pkg
